+++ apps/router-e2e/__e2e__/native-tabs/app/_layout.tsx
@@ -46,7 +46,11 @@ export default function Layout() {
         // indicatorColor="black"
         // sidebarAdaptable
         >
-          <NativeTabs.Trigger name="index">
+          <NativeTabs.Trigger
+            name="index"
+            contentStyle={{
+              experimental_backgroundImage: 'linear-gradient(135deg, #fbc2eb, #a6c1ee)',
+            }}>
             <NativeTabs.Trigger.Label
             // selectedStyle={{ color: '#0f0' }}
             >