+++ apps/router-e2e/__e2e__/native-tabs/app/explore/index.tsx
@@ -4,11 +4,10 @@ import { ScrollView } from 'react-native';
 import { Post } from '../../components/Post';
 
 export default function Index() {
-  const { colors } = useTheme();
   return (
     <ScrollView
       contentInsetAdjustmentBehavior="automatic"
-      style={{ flex: 1, backgroundColor: colors.background }}
+      style={{ flex: 1 }}
       contentContainerStyle={{
         padding: 32,
         gap: 16,