+++ apps/router-e2e/__e2e__/native-tabs/app/explore/news/[title].tsx
@@ -4,11 +4,10 @@ import { ScrollView, Text } from 'react-native';
 
 export default function Index() {
   const { title } = useLocalSearchParams();
-  const { colors } = useTheme();
   return (
     <ScrollView
       contentInsetAdjustmentBehavior="automatic"
-      style={{ flex: 1, backgroundColor: colors.background }}
+      style={{ flex: 1 }}
       contentContainerStyle={{
         padding: 32,
         gap: 16,