+++ apps/router-e2e/__e2e__/native-tabs/app/faces/index.tsx
@@ -1,14 +1,12 @@
-import { useTheme } from '@react-navigation/native';
 import { ScrollView } from 'react-native';
 
 import { Faces } from '../../components/faces';
 
 export default function Index() {
-  const { colors } = useTheme();
   return (
     <ScrollView
       contentInsetAdjustmentBehavior="automatic"
-      style={{ flex: 1, backgroundColor: colors.background }}
+      style={{ flex: 1 }}
       contentContainerStyle={{
         alignItems: 'center',
         padding: 32,