+++ apps/router-e2e/__e2e__/native-tabs/app/index.tsx
@@ -22,7 +22,7 @@ export default function Index() {
   return (
     <ScrollView
       contentInsetAdjustmentBehavior="automatic"
-      style={{ flex: 1, backgroundColor: colors.background }}
+      style={{ flex: 1 }}
       contentContainerStyle={{
         justifyContent: 'center',
         // alignItems: 'center',