+++ packages/expo-router/src/native-tabs/types.ts
@@ -1,6 +1,12 @@
 import type { DefaultRouterOptions } from '@react-navigation/native';
 import type { PropsWithChildren } from 'react';
-import type { ColorValue, ImageSourcePropType, StyleProp, TextStyle } from 'react-native';
+import type {
+  ColorValue,
+  ImageSourcePropType,
+  StyleProp,
+  TextStyle,
+  ViewStyle,
+} from 'react-native';
 import type { BottomTabsScreenProps } from 'react-native-screens';
 import type { SFSymbol } from 'sf-symbols-typescript';
 
@@ -35,6 +41,31 @@ export interface NativeTabOptions extends DefaultRouterOptions {
   hidden?: boolean;
   specialEffects?: BottomTabsScreenProps['specialEffects'];
   nativeProps?: NativeScreenProps;
+  contentStyle?: Pick<
+    ViewStyle,
+    | 'backgroundColor'
+    | 'experimental_backgroundImage'
+    | 'padding'
+    | 'paddingTop'
+    | 'paddingBottom'
+    | 'paddingLeft'
+    | 'paddingRight'
+    | 'paddingBlock'
+    | 'paddingBlockEnd'
+    | 'paddingBlockStart'
+    | 'paddingInline'
+    | 'paddingInlineEnd'
+    | 'paddingInlineStart'
+    | 'paddingEnd'
+    | 'paddingHorizontal'
+    | 'paddingVertical'
+    | 'paddingStart'
+    | 'alignContent'
+    | 'alignItems'
+    | 'justifyContent'
+    | 'flexDirection'
+    | 'gap'
+  >;
 }
 
 export type SymbolOrImageSource =
@@ -335,6 +366,12 @@ export interface NativeTabTriggerProps {
    * @platform ios
    */
   role?: NativeTabsTabBarItemRole;
+  /**
+   * The style applied to the content of the tab
+   *
+   * Note: Only certain style properties are supported.
+   */
+  contentStyle?: NativeTabOptions['contentStyle'];
 }
 
 const SUPPORTED_TAB_BAR_ITEM_ROLES = [